Stdout/Stderr/Return code reporting working
Note that this makes an important change. Instead of using S3, we are now
using the local filesystem for reporting back stdout/stderr/return code
to Cromwell. This is intended to be reverted in a future commit.

Also, the writing of files is at a non-standard place in the
Cromwell lifecycle. This will also be investigated later